APM-270 Create Patient Searcher
