+++ sandbox/routes/patient/get-patient.js
@@ -5,8 +5,6 @@ const nhsNumberHelper = require('../../helpers/nhs-number-helper')
 const dateValidator = require('../../validators/date-validator')
 
 const EXAMPLE_PATIENT_SMITH = JSON.parse(fs.readFileSync('mocks/Patient-Jane-Smith.json'))
-const EXAMPLE_PATIENT_SMYTHE = JSON.parse(fs.readFileSync('mocks/Patient-Jayne-Smyth.json'))
-
 
 module.exports = [
     /* Patient Search
@@ -34,23 +32,13 @@ module.exports = [
         path: '/Patient',
         handler: (request) => {
             
-            // TODO: Spilt down methods into seperate files
+            let patientSearcher = require("../../services/patient-searcher").init(request)
 
-            // TODO: This can be provided to a PatientSearcher to use to implement a more
-            // 'proper' search
-            const searchMap = {
-                "_exact-match": true,
-                "_history": true,
-                "_fuzzy-match": true,
-                "_max-results": true,
-                family: '$.name[?(@.use="usual")].family', // Usual family name
-                given: true,
-                gender: true,
-                birthdate: true,
-                "death-date": true,
-                "address-postcode": true,
-                organisation: true,
-            };
+            if (!patientSearcher.requestContainsParameters()) {
+                throw Boom.badRequest(
+                    "Not enough search parameters were provided to be able to make a search",
+                    {operationOutcomeCode: "required", apiErrorCode: "tooFewSearchParams"})
+            }
 
             // If provided, validate birthdate, death-date params
             // TODO: birthdate range
@@ -63,104 +51,7 @@ module.exports = [
 
             });
 
-            // Check for too few search params
-            // TODO: Improve this - currently checks for *any* search param
-            let hasAnySearchParam = false
-            for (let p of Object.keys(searchMap)) {
-                if (request.query[p]) {
-                    hasAnySearchParam = true
-                    break
-                }
-            }
-            if (!hasAnySearchParam) {
-                throw Boom.badRequest(
-                    "Not enough search parameters were provided to be able to make a search",
-                    {operationOutcomeCode: "required", apiErrorCode: "tooFewSearchParams"})
-            }
-
-            // Build our empty search response
-            let response = {
-                resourceType: "Bundle",
-                type: "searchset",
-                timestamp: Date.now(),
-                total: 0,
-                entry: []
-            }
-
-            // Build a response with one patient
-            let examplePatientResponse = function() {
-                response.total = 1
-                response.entry.push({
-                    search: {
-                        score: 1.0
-                    },
-                    resource: EXAMPLE_PATIENT_SMITH,
-                })
-                return response
-            }
-
-            // Build example response with 2 patients
-            let exampleResponseWithTwoPatients = function () {
-                response.total = 2
-                response.entry.push({
-                    search: {
-                        score: 0.8343
-                    },
-                        resource: [EXAMPLE_PATIENT_SMITH, EXAMPLE_PATIENT_SMYTHE]
-                })
-                return response
-            }
-
-            // Can be moved elsewhere
-            let containsSearchParameters = function(searchParameters) {
-                for (let p of Object.keys(searchParameters)) {
-                    if (!request.query[p] || request.query[p].toLowerCase() !== searchParameters[p].toLowerCase()) {
-                        return false
-                    }
-                }
-                return true
-            }
-
-
-            // Check if wildcard provided
-            const  wildcardSearchParams = {
-                family: "Sm*",
-                gender: "female",
-                birthdate: "2010-10-22"
-            }
-            let wildcardMatch = containsSearchParameters(wildcardSearchParams)
-
-            // Perform a search with max result set using the wildcard params and the max-result parameter
-            if (wildcardMatch && request.query["_max-results"]) {
-                if (isNaN(request.query["_max-results"]) || request.query["_max-results"] < 1) {
-                    // Happy path only (For now)
-                    throw Boom.badRequest("TBC", {
-                        operationOutcomeCode: "TBC", apiErrorCode: "TBC"
-                    })
-                } else if (request.query["_max-results"] == 1) {
-                    return examplePatientResponse()
-                } else {
-                    return exampleResponseWithTwoPatients()
-                }
-            // Perform a advanced search as wildcard provided and max-result parameter not set
-            } else if (wildcardMatch) {
-                return exampleResponseWithTwoPatients()
-            }
-
-
-            // Perform a 'simple search'
-            const simpleSearchParams = {
-                family: "Smith",
-                gender: "female",
-                birthdate: "2010-10-22",
-            }
-            let simpleMatch = containsSearchParameters(simpleSearchParams)
-            // If so, try it
-            if (simpleMatch) {
-                return examplePatientResponse()
-            }
-
-            return response
+            return patientSearcher.search()
         }
     }
 ]
